--- 02含批阅意见的实验报告/第10小组：题目名称/源程序及数据文件/租借信息管理系统.plg
@@ -1,32 +1,0 @@
-<html>
-<body>
-<pre>
-<h1>Build Log</h1>
-<h3>
---------------------Configuration: 1 - Win32 Debug--------------------
-</h3>
-<h3>Command Lines</h3>
-Creating temporary file "C:\Users\ADMINI~1\AppData\Local\Temp\RSPDF0.tmp" with contents
-[
-/nologo /MLd /W3 /Gm /GX /ZI /Od /D "WIN32" /D "_DEBUG" /D "_CONSOLE" /D "_MBCS" /Fp"Debug/1.pch" /YX /Fo"Debug/" /Fd"Debug/" /FD /GZ /c 
-"E:\1.c"
-]
-Creating command line "cl.exe @C:\Users\ADMINI~1\AppData\Local\Temp\RSPDF0.tmp" 
-Creating temporary file "C:\Users\ADMINI~1\AppData\Local\Temp\RSPDF1.tmp" with contents
-[
-kernel32.lib user32.lib gdi32.lib winspool.lib comdlg32.lib advapi32.lib shell32.lib ole32.lib oleaut32.lib uuid.lib odbc32.lib odbccp32.lib /nologo /subsystem:console /incremental:yes /pdb:"Debug/1.pdb" /debug /machine:I386 /out:"Debug/1.exe" /pdbtype:sept 
-.\Debug\1.obj
-]
-Creating command line "link.exe @C:\Users\ADMINI~1\AppData\Local\Temp\RSPDF1.tmp"
-<h3>Output Window</h3>
-Compiling...
-1.c
-Linking...
-
-
-
-<h3>Results</h3>
-1.exe - 0 error(s), 0 warning(s)
-</pre>
-</body>
-</html>
