+++ 02含批阅意见的实验报告/第10小组：题目名称/源程序及数据文件/租车信息系统.dsp
@@ -1,24 +1,24 @@
-# Microsoft Developer Studio Project File - Name="1" - Package Owner=<4>
+# Microsoft Developer Studio Project File - Name="租车信息系统" - Package Owner=<4>
 # Microsoft Developer Studio Generated Build File, Format Version 6.00
 # ** DO NOT EDIT **
 
 # TARGTYPE "Win32 (x86) Console Application" 0x0103
 
-CFG=1 - Win32 Debug
+CFG=租车信息系统 - Win32 Debug
 !MESSAGE This is not a valid makefile. To build this project using NMAKE,
 !MESSAGE use the Export Makefile command and run
 !MESSAGE 
-!MESSAGE NMAKE /f "1.mak".
+!MESSAGE NMAKE /f "租车信息系统.mak".
 !MESSAGE 
 !MESSAGE You can specify a configuration when running NMAKE
 !MESSAGE by defining the macro CFG on the command line. For example:
 !MESSAGE 
-!MESSAGE NMAKE /f "1.mak" CFG="1 - Win32 Debug"
+!MESSAGE NMAKE /f "租车信息系统.mak" CFG="租车信息系统 - Win32 Debug"
 !MESSAGE 
 !MESSAGE Possible choices for configuration are:
 !MESSAGE 
-!MESSAGE "1 - Win32 Release" (based on "Win32 (x86) Console Application")
-!MESSAGE "1 - Win32 Debug" (based on "Win32 (x86) Console Application")
+!MESSAGE "租车信息系统 - Win32 Release" (based on "Win32 (x86) Console Application")
+!MESSAGE "租车信息系统 - Win32 Debug" (based on "Win32 (x86) Console Application")
 !MESSAGE 
 
 # Begin Project
@@ -28,7 +28,7 @@ CFG=1 - Win32 Debug
 CPP=cl.exe
 RSC=rc.exe
 
-!IF  "$(CFG)" == "1 - Win32 Release"
+!IF  "$(CFG)" == "租车信息系统 - Win32 Release"
 
 # PROP BASE Use_MFC 0
 # PROP BASE Use_Debug_Libraries 0
@@ -51,7 +51,7 @@ LINK32=link.exe
 # ADD BASE LINK32 kernel32.lib user32.lib gdi32.lib winspool.lib comdlg32.lib advapi32.lib shell32.lib ole32.lib oleaut32.lib uuid.lib odbc32.lib odbccp32.lib /nologo /subsystem:console /machine:I386
 # ADD LINK32 kernel32.lib user32.lib gdi32.lib winspool.lib comdlg32.lib advapi32.lib shell32.lib ole32.lib oleaut32.lib uuid.lib odbc32.lib odbccp32.lib /nologo /subsystem:console /machine:I386
 
-!ELSEIF  "$(CFG)" == "1 - Win32 Debug"
+!ELSEIF  "$(CFG)" == "租车信息系统 - Win32 Debug"
 
 # PROP BASE Use_MFC 0
 # PROP BASE Use_Debug_Libraries 1
@@ -78,11 +78,11 @@ LINK32=link.exe
 
 # Begin Target
 
-# Name "1 - Win32 Release"
-# Name "1 - Win32 Debug"
+# Name "租车信息系统 - Win32 Release"
+# Name "租车信息系统 - Win32 Debug"
 # Begin Source File
 
-SOURCE=.\1.c
+SOURCE=".\租车信息系统.cpp"
 # End Source File
 # End Target
 # End Project
