+++ 02含批阅意见的实验报告/第10小组：题目名称/源程序及数据文件/租车信息系统.dsw
@@ -1,9 +1,9 @@
 Microsoft Developer Studio Workspace File, Format Version 6.00
-# WARNING: DO NOT EDIT OR DELETE THIS WORKSPACE FILE!
+# 警告: 不能编辑或删除该工作区文件！
 
 ###############################################################################
 
-Project: "1"=.\1.dsp - Package Owner=<4>
+Project: "租车信息系统"=".\租车信息系统.dsp" - Package Owner=<4>
 
 Package=<5>
 {{{
