+++ 02含批阅意见的实验报告/第10小组：题目名称/源程序及数据文件/租车信息系统.plg
@@ -1,0 +1,24 @@
+<html>
+<body>
+<pre>
+<h1>Build Log</h1>
+<h3>
+--------------------Configuration: 租车信息系统 - Win32 Debug--------------------
+</h3>
+<h3>Command Lines</h3>
+Creating temporary file "C:\Users\ADMINI~1\AppData\Local\Temp\RSPFA9A.tmp" with contents
+[
+kernel32.lib user32.lib gdi32.lib winspool.lib comdlg32.lib advapi32.lib shell32.lib ole32.lib oleaut32.lib uuid.lib odbc32.lib odbccp32.lib /nologo /subsystem:console /incremental:yes /pdb:"Debug/租车信息系统.pdb" /debug /machine:I386 /out:"Debug/租车信息系统.exe" /pdbtype:sept 
+".\Debug\租车信息系统.obj"
+]
+Creating command line "link.exe @C:\Users\ADMINI~1\AppData\Local\Temp\RSPFA9A.tmp"
+<h3>Output Window</h3>
+Linking...
+
+
+
+<h3>Results</h3>
+租车信息系统.exe - 0 error(s), 0 warning(s)
+</pre>
+</body>
+</html>
